requirements.txt
torch>2.4
torchvision
torchaudio
numpy<2
librosa
matplotlib
tensorboardX
slune-lib @ git+https://github.com/h-0-0/slune@main
gdown
tensorflow_datasets
tensorflow
torchfile
datasets
h5py
pillow
accelerate
torchmetrics